All Score Bar Implemented

+++ 2D-Game/Super-Mario/Super_Mario/Score.h
@@ -21,7 +21,7 @@ private:
 	int score;
 	int coins;
 	pair<int, int> world;
-	int time;
+	double time;
 	int lives;
 	Text text;
